Reorganize unit tests based on discussion with Jim Edwards
My original idea had been to have a 'tests' subdirectory of each
production code directory, and to have unit tests live there. However,
Jim Edwards suggested having a top-level tests directory, and then
having a directory structure under that that mirrors the production code
directory structure. This commit does that.

Note that I have also moved case_fake.py to a higher level within the
tests directory, mirroring the location of case.py.

Test suite: python -m unittest discover -s CIME/tests -t .
Test baseline: N/A
Test namelist changes: N/A
Test status: pass

Fixes: None

User interface changes?: No

Code review: None
